[launcher] Remove writing default service TTL to config.json
In current design, if TTL field is not defined in initial config.json,
default value will be written to config.json on install and then will be
parsed to be put into the DB. In this patch default value is not written
to config.json but directly put into DB. If TTL field exists in initial
config.json, this value is placed to DB.

+++ config/config.go
@@ -54,7 +54,7 @@ type Config struct {
 	ServiceDiscoveryURL string     `json:"serviceDiscovery"`
 	VISServerURL        string     `json:"visServer"`
 	WorkingDir          string     `json:"workingDir"`
-	DefaultServiceTTL   uint       `json:"defaultServiceTTLDays"`
+	DefaultServiceTTL   uint64     `json:"defaultServiceTTLDays"`
 	Monitoring          Monitoring `json:"monitoring"`
 }
